estableco las tablas relacionadas

+++ version_3/mainrelationships.py
@@ -3,7 +3,7 @@
 from sqlalchemy import Column, Integer, String, ForeignKey, Sequence, create_engine
 from sqlalchemy.orm import declarative_base, relationship, sessionmaker
 
-engine  = create_engine('sqlite:///orm.db')
+engine  = create_engine('sqlite:///orm_relationships.db')
 
 Session = sessionmaker( bind=engine)
 session = Session()
@@ -12,27 +12,39 @@ Base = declarative_base()
 
 class User(Base):
     __tablename__ = 'users'
-    #id = column(Integer, Sequence('user_id_seq'), primary_key=True)
     id = Column(Integer, Sequence('user_id_seq'), primary_key=True)
     name = Column(String(50))
     email = Column(String(50))
 
+    posts = relationship('Post', back_populates='user')
+
+class Post(Base):
+    __tablename__ = 'posts'
+    id = Column(Integer, Sequence('post_id_seq'), primary_key=True)
+    title = Column(String(100))
+    content = Column(String(500))
+    user_id = Column(Integer, ForeignKey('users.id'))
+    user = relationship('User', back_populates='posts')
+
+    user = relationship('User', back_populates='posts')
+
 Base.metadata.create_all(engine)
 
-#user1 = User(name='Alice', email='alice@example.com')
-#user2 = User(name='Bob', email=' bob@example.com')
+user1 = User(name='Alice', email='alice@example.com')
+user2 = User(name='Bob', email='bob@example.com')
 
 #insert data:
-#session.add_all([user1, user2])
-#session.commit()
+session.add_all([user1, user2])
+session.commit()
 
 #query data:
 user = session.query(User).filter_by(name=('Alice')).first()
-print(f'User: {user.name}, Email: {user.email}')
+if user is not None:
+    print(f'User: {user.name}, Email: {user.email}')
 
 #delete data:
-session.delete(user)
-session.commit()
+    session.delete(user)
+    session.commit()
 
 #verify deletion
 user = session.query(User).filter_by(name='Alice').first()
@@ -50,6 +62,8 @@ session.commit()
 
 #verify update
 user = session.query(User).filter_by(name='Bob').first()
-print(f'User: {user.name}, Email: {user.email}')
+if user is not None:
+    print(f'User: {user.name}, Email: {user.email}')
 
 
+session.close()
